change the function to type handle and move the functions to a new file
