+++ contract/system/htlc_test.go
@@ -1,3 +1,8 @@
+/**
+*  @file
+*  @copyright defined in go-seele/LICENSE
+ */
+
 package system
 
 import (
